requirements.txt

# Core Flask Dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.1.1
WTForms==3.0.1

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# HTTP and API
requests==2.31.0
urllib3==2.0.4

# PDF Processing (lightweight alternatives)
PyPDF2==3.0.1
python-docx==0.8.11

# Image Processing (lightweight)
Pillow==10.0.1

# Security and Encryption
cryptography==41.0.4
itsdangerous==2.1.2

# Payment Processing
stripe==6.6.0

# WSGI Server
gunicorn==21.2.0

# Utilities
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.6.2
